requirements.txt
tensorflow==1.12.0
opencv-python==3.4.4.19 
numpy==1.15.4
Cython==0.29.1
h5py==2.9.0
Pillow==5.4.1
pytesseract==0.2.6
